# target/classes/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: API de Gestão de Portfólio
  description: Contrato de API para o CMS e exibição da Landing Page do Portfólio (Modulo Portfolio)
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Servidor Local de Desenvolvimento

paths:
  /api/v1/portfolio/perfil:
    get:
      summary: Obter o perfil profissional público
      description: Retorna as informações do perfil profissional cadastrado na landing page.
      responses:
        '200':
          description: Detalhes do perfil retornados com sucesso.
          headers:
            X-Trace-Id:
              schema:
                type: string
              description: Identificador de rastreabilidade da requisição.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PerfilResponse'
        '404':
          description: Perfil não configurado.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'

  /api/v1/portfolio/projetos:
    get:
      summary: Listar projetos desenvolvidos
      description: Retorna a lista de projetos associados com suas respectivas tecnologias ordanados por data de desenvolvimento decrescente.
      responses:
        '200':
          description: Lista de projetos retornada com sucesso.
          headers:
            X-Trace-Id:
              schema:
                type: string
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProjetoResponse'

  /api/v1/admin/portfolio/projetos:
    post:
      summary: Criar novo projeto (Administrativo)
      description: Cadastra um novo projeto e associa com as tecnologias informadas. Requer autorização.
      parameters:
        - name: Idempotency-Key
          in: header
          required: true
          schema:
            type: string
          description: Chave única de idempotência para evitar requisições duplicadas.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjetoRequest'
      responses:
        '201':
          description: Projeto criado com sucesso.
          headers:
            Location:
              schema:
                type: string
              description: URI de acesso ao novo projeto criado.
            X-Trace-Id:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjetoResponse'
        '400':
          description: Dados de entrada inválidos ou Idempotency-Key ausente.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '409':
          description: Conflito de Idempotência (Chave em processamento ou já executada).
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'

  /api/v1/admin/portfolio/projetos/{id}:
    put:
      summary: Atualizar projeto existente (Administrativo)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjetoRequest'
      responses:
        '200':
          description: Projeto atualizado com sucesso.
          headers:
            X-Trace-Id:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjetoResponse'
        '404':
          description: Projeto ou tecnologia associada não encontrados.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '409':
          description: Conflito de Concorrência Otimista (versão inconsistente).
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'

    delete:
      summary: Excluir projeto (Administrativo)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Projeto excluído com sucesso.
          headers:
            X-Trace-Id:
              schema:
                type: string
        '404':
          description: Projeto não encontrado.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'

  /api/v1/admin/portfolio/tecnologias:
    post:
      summary: Criar nova tecnologia (Administrativo)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TecnologiaRequest'
      responses:
        '201':
          description: Tecnologia criada com sucesso.
          headers:
            Location:
              schema:
                type: string
            X-Trace-Id:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TecnologiaResponse'
        '400':
          description: Nome duplicado ou dados inválidos.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'

components:
  schemas:
    PerfilResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        nome:
          type: string
        titulo:
          type: string
        bio:
          type: string
        linkLinkedin:
          type: string
        linkGithub:
          type: string
        urlFoto:
          type: string
        usuarioId:
          type: string
          format: uuid
        versao:
          type: integer

    ProjetoRequest:
      type: object
      required:
        - titulo
      properties:
        titulo:
          type: string
        descricao:
          type: string
        urlCapa:
          type: string
        linkProducao:
          type: string
        linkRepositorio:
          type: string
        dataDesenvolvimento:
          type: string
          format: date
        tecnologiaIds:
          type: array
          items:
            type: string
            format: uuid

    ProjetoResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        titulo:
          type: string
        descricao:
          type: string
        urlCapa:
          type: string
        linkProducao:
          type: string
        linkRepositorio:
          type: string
        dataDesenvolvimento:
          type: string
          format: date
        tecnologias:
          type: array
          items:
            $ref: '#/components/schemas/TecnologiaResponse'
        versao:
          type: integer

    TecnologiaRequest:
      type: object
      required:
        - nome
      properties:
        nome:
          type: string
        urlIcone:
          type: string

    TecnologiaResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        nome:
          type: string
        urlIcone:
          type: string
        versao:
          type: integer

    ProblemDetails:
      type: object
      required:
        - status
        - title
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
          format: uri
        timestamp:
          type: string
          format: date-time